+++ packages/imperative/src/DefinitionTreeResolver.ts
@@ -47,6 +47,11 @@ export class DefinitionTreeResolver {
                 "configuration."
             });
         }
+        if (childrenDefinitions == null) {
+            childrenDefinitions = [];
+        } else if (childrenModuleGlobs == null) {
+            childrenModuleGlobs = [];
+        }
         const root: ICommandDefinition =
             {
                 name: "",
@@ -90,7 +95,7 @@ export class DefinitionTreeResolver {
      */
     public static combineAllCmdDefs(
         callerDir: string,
-        cmdDefs: ICommandDefinition[]= [],
+        cmdDefs: ICommandDefinition[] = [],
         cmdModuleGlobs: string[] = [],
         addBaseProfile?: boolean
     ): ICommandDefinition[] {